Refactor components (#44)
* Split the page into components

* Change css

* fix bug due to css

---------
