+++ src/app/globals.css
@@ -2,6 +2,7 @@
 @source '../**/*.{js,ts,jsx,tsx}';
 
 @import 'tw-animate-css';
+@import '../styles/components.css';
 
 @custom-variant dark (&:is(.dark *));
